looking for a working implementation online (two implementations do not work so far)

+++ src/solar-test.c
@@ -43,7 +43,7 @@ static void printSunrise2(
 	#pragma GCC diagnostic pop
     double hours;
     double minutes = modf (localT, &hours)*60;
-    printf("%2.0g:%2.0g\n",hours,minutes);
+    printf("%02d:%02d\n", (int) hours, (int) minutes);
 }
 
 __attribute__ ((nothrow, warn_unused_result))
